cascade delete time, no more state crash
