+++ frontend/src/components/EditPropertyForm/EditPropertyForm.js
@@ -1,6 +1,6 @@
 import './EditPropertyForm.css'
 import { useState } from 'react';
-import { changeProperty, deleteProperty } from '../../store/properties';
+import { changeProperty, deleteProperty } from '../../store/property';
 import { useDispatch, useSelector } from 'react-redux';
 import { useHistory } from 'react-router-dom'
 import { Redirect } from 'react-router-dom';
